backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1

# PDF
PyMuPDF==1.24.14

# Tokenisation
tiktoken==0.8.0

# LLM
anthropic==0.40.0

# Embeddings
sentence-transformers==3.3.1

# Vector store
faiss-cpu==1.9.0

# Graph DB
neo4j==5.27.0

# Numerics
numpy==1.26.4
